+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Campaign Changelog
 
+## Unreleased
+- Preflight sendout checks
+- Changed "Settings" to "Plugin Settings"
+
 ## 1.0.0-beta1 - 2018-03-07
 - Initial release of public beta